fix(desktop): address PR review feedback on tasks integration
- SafeImage: drop http://, allow only data: and https:// (cleartext is
  no longer needed; GitHub user-attachments are https).
- ElectronTRPCProvider: compose shouldDehydrateQuery with
  defaultShouldDehydrateQuery so pending/error queries are excluded
  from the persisted IndexedDB cache.
- search-pull-requests / search-github-issues: direct-lookup branches
  now echo the requested page instead of hardcoding 1, matching the
  search branches.
- PullRequestsContent / GitHubIssuesContent: short-circuit row keydown
  when the event target isn't the row itself, so Enter/Space on a
  nested action button no longer also navigates to the preview.
  repoMismatch text gets select-text cursor-text so the repo string
  is copyable (per AGENTS.md).
- AssigneeFilter / ProjectFilter / StatusFilter: add aria-label to the
  icon-only popover triggers (title alone is tooltip-only on many
  screen readers).
- Bump @tanstack/react-query to ^5.100.9 to satisfy
  @tanstack/react-query-persist-client peer requirement.
- Extract shared normalizePRState helper next to the PRState type so
  the PR list and detail page can't drift.
