+++ apps/desktop/package.json
@@ -105,7 +105,7 @@
 		"@tanstack/node-db-sqlite-persistence": "0.1.9",
 		"@tanstack/query-async-storage-persister": "^5.100.9",
 		"@tanstack/react-db": "0.1.83",
-		"@tanstack/react-query": "^5.90.19",
+		"@tanstack/react-query": "^5.100.9",
 		"@tanstack/react-query-persist-client": "^5.100.9",
 		"@tanstack/react-router": "^1.147.3",
 		"@tanstack/react-table": "^8.21.3",
