+++ apps/desktop/src/renderer/screens/main/components/PRIcon/index.ts
@@ -1,1 +1,2 @@
+export { normalizePRState } from "./normalizePRState";
 export { PRIcon, type PRState } from "./PRIcon";
